move node account from x/crosschain to x/observer
