+++ cmd/zetacored/observer_accounts.go
@@ -69,7 +69,7 @@ func AddObserverAccountsCmd() *cobra.Command {
 			}
 			var observerMapper []*types.ObserverMapper
 			var grantAuthorizations []authz.GrantAuthorization
-			var nodeAccounts []*crosschaintypes.NodeAccount
+			var nodeAccounts []*types.NodeAccount
 			var keygenPubKeys []string
 			observersForChain := map[int64][]string{}
 			// DefaultChainsList is based on Build Flags
@@ -121,11 +121,11 @@ func AddObserverAccountsCmd() *cobra.Command {
 						Secp256k1: pubkey,
 						Ed25519:   "",
 					}
-					na := crosschaintypes.NodeAccount{
+					na := types.NodeAccount{
 						Operator:       info.ObserverAddress,
 						GranteeAddress: info.ZetaClientGranteeAddress,
 						GranteePubkey:  &pubkeySet,
-						NodeStatus:     crosschaintypes.NodeStatus_Active,
+						NodeStatus:     types.NodeStatus_Active,
 					}
 					nodeAccounts = append(nodeAccounts, &na)
 				}
@@ -156,7 +156,6 @@ func AddObserverAccountsCmd() *cobra.Command {
 
 			// Add node accounts to cross chain genesis state
 			zetaCrossChainGenState := crosschaintypes.GetGenesisStateFromAppState(cdc, appState)
-			zetaCrossChainGenState.NodeAccountList = nodeAccounts
 			keyGenStatus := crosschaintypes.KeygenStatus_PendingKeygen
 			if keyGenBlock == 0 {
 				keyGenStatus = crosschaintypes.KeygenStatus_KeyGenSuccess
@@ -185,6 +184,7 @@ func AddObserverAccountsCmd() *cobra.Command {
 			// Add observers to observer genesis state
 			zetaObserverGenState := types.GetGenesisStateFromAppState(cdc, appState)
 			zetaObserverGenState.Observers = observerMapper
+			zetaObserverGenState.NodeAccountList = nodeAccounts
 
 			// Add grant authorizations to authz genesis state
 			var authzGenState authz.GenesisState
