+++ docs/openapi/openapi.swagger.yaml
@@ -26997,7 +26997,7 @@ paths:
         "200":
           description: A successful response.
           schema:
-            $ref: '#/definitions/crosschainQueryAllNodeAccountResponse'
+            $ref: '#/definitions/observerQueryAllNodeAccountResponse'
         default:
           description: An unexpected error response.
           schema:
@@ -27056,7 +27056,7 @@ paths:
         "200":
           description: A successful response.
           schema:
-            $ref: '#/definitions/crosschainQueryGetNodeAccountResponse'
+            $ref: '#/definitions/observerQueryGetNodeAccountResponse'
         default:
           description: An unexpected error response.
           schema:
@@ -50099,8 +50099,6 @@ definitions:
     type: object
   crosschainMsgRemoveFromOutTxTrackerResponse:
     type: object
-  crosschainMsgSetNodeKeysResponse:
-    type: object
   crosschainMsgUpdateKeygenResponse:
     type: object
   crosschainMsgUpdatePermissionFlagsResponse:
@@ -50109,27 +50107,6 @@ definitions:
     type: object
   crosschainMsgVoteOnObservedOutboundTxResponse:
     type: object
-  crosschainNodeAccount:
-    type: object
-    properties:
-      operator:
-        type: string
-      granteeAddress:
-        type: string
-      granteePubkey:
-        $ref: '#/definitions/commonPubKeySet'
-      nodeStatus:
-        $ref: '#/definitions/crosschainNodeStatus'
-  crosschainNodeStatus:
-    type: string
-    enum:
-      - Unknown
-      - Whitelisted
-      - Standby
-      - Ready
-      - Active
-      - Disabled
-    default: Unknown
   crosschainOutTxTracker:
     type: object
     properties:
@@ -50257,16 +50234,6 @@ definitions:
           $ref: '#/definitions/crosschainLastBlockHeight'
       pagination:
         $ref: '#/definitions/v1beta1PageResponse'
-  crosschainQueryAllNodeAccountResponse:
-    type: object
-    properties:
-      NodeAccount:
-        type: array
-        items:
-          type: object
-          $ref: '#/definitions/crosschainNodeAccount'
-      pagination:
-        $ref: '#/definitions/v1beta1PageResponse'
   crosschainQueryAllOutTxTrackerByChainResponse:
     type: object
     properties:
@@ -50335,11 +50302,6 @@ definitions:
     properties:
       LastBlockHeight:
         $ref: '#/definitions/crosschainLastBlockHeight'
-  crosschainQueryGetNodeAccountResponse:
-    type: object
-    properties:
-      NodeAccount:
-        $ref: '#/definitions/crosschainNodeAccount'
   crosschainQueryGetOutTxTrackerResponse:
     type: object
     properties:
@@ -50535,6 +50497,27 @@ definitions:
     type: object
   observerMsgUpdateCoreParamsResponse:
     type: object
+  observerNodeAccount:
+    type: object
+    properties:
+      operator:
+        type: string
+      granteeAddress:
+        type: string
+      granteePubkey:
+        $ref: '#/definitions/commonPubKeySet'
+      nodeStatus:
+        $ref: '#/definitions/observerNodeStatus'
+  observerNodeStatus:
+    type: string
+    enum:
+      - Unknown
+      - Whitelisted
+      - Standby
+      - Ready
+      - Active
+      - Disabled
+    default: Unknown
   observerObservationType:
     type: string
     enum:
@@ -50575,6 +50558,16 @@ definitions:
       - out_tx_tracker
     default: stop_inbound_cctx
     title: '- stop_inbound_cctx: some observer sees inbound tx'
+  observerQueryAllNodeAccountResponse:
+    type: object
+    properties:
+      NodeAccount:
+        type: array
+        items:
+          type: object
+          $ref: '#/definitions/observerNodeAccount'
+      pagination:
+        $ref: '#/definitions/v1beta1PageResponse'
   observerQueryAllObserverMappersResponse:
     type: object
     properties:
@@ -50607,6 +50600,11 @@ definitions:
     properties:
       core_params:
         $ref: '#/definitions/observerCoreParamsList'
+  observerQueryGetNodeAccountResponse:
+    type: object
+    properties:
+      NodeAccount:
+        $ref: '#/definitions/observerNodeAccount'
   observerQueryObserversByChainResponse:
     type: object
     properties:
