+++ proto/observer/node_account.proto
@@ -1,7 +1,7 @@
 syntax = "proto3";
-package zetachain.zetacore.crosschain;
+package zetachain.zetacore.observer;
 
-option go_package = "github.com/zeta-chain/zetacore/x/crosschain/types";
+option go_package = "github.com/zeta-chain/zetacore/x/observer/types";
 
 import "gogoproto/gogo.proto";
 import "common/common.proto";
