+++ testutil/network/genesis_state.go
@@ -30,15 +30,14 @@ func SetupZetaGenesisState(t *testing.T, genesisState map[string]json.RawMessage
 	// Cross-chain genesis state
 	var crossChainGenesis types.GenesisState
 	assert.NoError(t, codec.UnmarshalJSON(genesisState[types.ModuleName], &crossChainGenesis))
-	nodeAccountList := make([]*types.NodeAccount, len(observerList))
+	nodeAccountList := make([]*observerTypes.NodeAccount, len(observerList))
 	for i, operator := range observerList {
-		nodeAccountList[i] = &types.NodeAccount{
+		nodeAccountList[i] = &observerTypes.NodeAccount{
 			Operator:   operator,
-			NodeStatus: types.NodeStatus_Active,
+			NodeStatus: observerTypes.NodeStatus_Active,
 		}
 	}
 
-	crossChainGenesis.NodeAccountList = nodeAccountList
 	crossChainGenesis.Keygen = &types.Keygen{
 		Status:         types.KeygenStatus_PendingKeygen,
 		GranteePubkeys: observerList,
@@ -74,6 +73,7 @@ func SetupZetaGenesisState(t *testing.T, genesisState map[string]json.RawMessage
 		}
 	}
 	observerGenesis.Observers = observerMapper
+	observerGenesis.NodeAccountList = nodeAccountList
 	observerGenesisBz, err := codec.MarshalJSON(&observerGenesis)
 	assert.NoError(t, err)
