+++ x/crosschain/client/cli/tx.go
@@ -35,7 +35,6 @@ func GetTxCmd() *cobra.Command {
 	cmd.AddCommand(CmdNonceVoter())
 	cmd.AddCommand(CmdCCTXOutboundVoter())
 	cmd.AddCommand(CmdCCTXInboundVoter())
-	cmd.AddCommand(CmdSetNodeKeys())
 	cmd.AddCommand(CmdRemoveFromWatchList())
 	cmd.AddCommand(CmdUpdatePermissionFlags())
 	cmd.AddCommand(CmdUpdateKeygen())
