+++ x/crosschain/client/querytests/chain_nonces.go
@@ -13,7 +13,7 @@ import (
 
 func (s *CliTestSuite) TestShowChainNonces() {
 	ctx := s.network.Validators[0].ClientCtx
-	objs := s.state.ChainNoncesList
+	objs := s.crossChainState.ChainNoncesList
 	common := []string{
 		fmt.Sprintf("--%s=json", tmcli.OutputFlag),
 	}
@@ -59,7 +59,7 @@ func (s *CliTestSuite) TestShowChainNonces() {
 
 func (s *CliTestSuite) TestListChainNonces() {
 	ctx := s.network.Validators[0].ClientCtx
-	objs := s.state.ChainNoncesList
+	objs := s.crossChainState.ChainNoncesList
 	request := func(next []byte, offset, limit uint64, total bool) []string {
 		args := []string{
 			fmt.Sprintf("--%s=json", tmcli.OutputFlag),
