+++ x/crosschain/client/querytests/out_tx_tracker.go
@@ -12,7 +12,7 @@ import (
 
 func (s *CliTestSuite) TestListOutTxTracker() {
 	ctx := s.network.Validators[0].ClientCtx
-	objs := s.state.OutTxTrackerList
+	objs := s.crossChainState.OutTxTrackerList
 	request := func(next []byte, offset, limit uint64, total bool) []string {
 		args := []string{
 			fmt.Sprintf("--%s=json", tmcli.OutputFlag),
