+++ x/crosschain/client/querytests/suite.go
@@ -11,15 +11,17 @@ import (
 	"github.com/zeta-chain/zetacore/testutil/network"
 	"github.com/zeta-chain/zetacore/testutil/nullify"
 	"github.com/zeta-chain/zetacore/x/crosschain/types"
+	observerTypes "github.com/zeta-chain/zetacore/x/observer/types"
 	"strconv"
 )
 
 type CliTestSuite struct {
 	suite.Suite
 
-	cfg     network.Config
-	network *network.Network
-	state   *types.GenesisState
+	cfg             network.Config
+	network         *network.Network
+	crossChainState *types.GenesisState
+	observerState   *observerTypes.GenesisState
 }
 
 func NewCLITestSuite(cfg network.Config) *CliTestSuite {
@@ -46,6 +48,7 @@ func (s *CliTestSuite) SetupSuite() {
 	}
 	s.cfg.GenesisState = network.SetupZetaGenesisState(s.T(), s.cfg.GenesisState, s.cfg.Codec, observerList)
 	s.AddCrossChainData(2)
+	s.AddObserverData(2)
 	net, err := network.New(s.T(), app.NodeDir, s.cfg)
 	s.Assert().NoError(err)
 	s.network = net
@@ -58,7 +61,17 @@ func (s *CliTestSuite) TearDownSuite() {
 	s.T().Log("tearing down genesis test suite")
 	s.network.Cleanup()
 }
-
+func (s *CliTestSuite) AddObserverData(n int) {
+	state := observerTypes.GenesisState{}
+	s.Require().NoError(s.cfg.Codec.UnmarshalJSON(s.cfg.GenesisState[observerTypes.ModuleName], &state))
+	for i := 0; i < n; i++ {
+		state.NodeAccountList = append(state.NodeAccountList, &observerTypes.NodeAccount{Operator: strconv.Itoa(i), GranteeAddress: "signer"})
+	}
+	buf, err := s.cfg.Codec.MarshalJSON(&state)
+	s.Require().NoError(err)
+	s.cfg.GenesisState[observerTypes.ModuleName] = buf
+	s.observerState = &state
+}
 func (s *CliTestSuite) AddCrossChainData(n int) {
 	state := types.GenesisState{}
 	s.Require().NoError(s.cfg.Codec.UnmarshalJSON(s.cfg.GenesisState[types.ModuleName], &state))
@@ -95,9 +108,6 @@ func (s *CliTestSuite) AddCrossChainData(n int) {
 		KeyGenZetaHeight:    1,
 	}
 	for i := 0; i < n; i++ {
-		state.NodeAccountList = append(state.NodeAccountList, &types.NodeAccount{Operator: strconv.Itoa(i), GranteeAddress: "signer"})
-	}
-	for i := 0; i < n; i++ {
 		outTxTracker := types.OutTxTracker{
 			Index:   fmt.Sprintf("%d-%d", i, i),
 			ChainId: int64(i),
@@ -120,5 +130,5 @@ func (s *CliTestSuite) AddCrossChainData(n int) {
 	buf, err := s.cfg.Codec.MarshalJSON(&state)
 	s.Require().NoError(err)
 	s.cfg.GenesisState[types.ModuleName] = buf
-	s.state = &state
+	s.crossChainState = &state
 }
