+++ x/crosschain/client/querytests/tss.go
@@ -11,7 +11,7 @@ import (
 
 func (s *CliTestSuite) TestShowTSS() {
 	ctx := s.network.Validators[0].ClientCtx
-	obj := s.state.Tss
+	obj := s.crossChainState.Tss
 	common := []string{
 		fmt.Sprintf("--%s=json", tmcli.OutputFlag),
 	}
