+++ x/crosschain/keeper/keeper_utils.go
@@ -51,7 +51,7 @@ func (k Keeper) IsAuthorized(ctx sdk.Context, address string, chain *common.Chai
 
 // IsAuthorized checks whether a signer is authorized to sign , by checking their address against the observer mapper which contains the observer list for the chain and type
 func (k Keeper) IsAuthorizedNodeAccount(ctx sdk.Context, address string) bool {
-	_, found := k.GetNodeAccount(ctx, address)
+	_, found := k.zetaObserverKeeper.GetNodeAccount(ctx, address)
 	if found {
 		return true
 	}
