+++ x/observer/client/cli/cli_node_account.go
@@ -1,0 +1,71 @@
+package cli
+
+import (
+	"context"
+
+	"github.com/cosmos/cosmos-sdk/client"
+	"github.com/cosmos/cosmos-sdk/client/flags"
+	"github.com/spf13/cobra"
+	"github.com/zeta-chain/zetacore/x/observer/types"
+)
+
+func CmdListNodeAccount() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "list-node-account",
+		Short: "list all NodeAccount",
+		RunE: func(cmd *cobra.Command, args []string) error {
+			clientCtx := client.GetClientContextFromCmd(cmd)
+
+			pageReq, err := client.ReadPageRequest(cmd.Flags())
+			if err != nil {
+				return err
+			}
+
+			queryClient := types.NewQueryClient(clientCtx)
+
+			params := &types.QueryAllNodeAccountRequest{
+				Pagination: pageReq,
+			}
+
+			res, err := queryClient.NodeAccountAll(context.Background(), params)
+			if err != nil {
+				return err
+			}
+
+			return clientCtx.PrintProto(res)
+		},
+	}
+
+	flags.AddPaginationFlagsToCmd(cmd, cmd.Use)
+	flags.AddQueryFlagsToCmd(cmd)
+
+	return cmd
+}
+
+func CmdShowNodeAccount() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "show-node-account [index]",
+		Short: "shows a NodeAccount",
+		Args:  cobra.ExactArgs(1),
+		RunE: func(cmd *cobra.Command, args []string) error {
+			clientCtx := client.GetClientContextFromCmd(cmd)
+
+			queryClient := types.NewQueryClient(clientCtx)
+
+			params := &types.QueryGetNodeAccountRequest{
+				Index: args[0],
+			}
+
+			res, err := queryClient.NodeAccount(context.Background(), params)
+			if err != nil {
+				return err
+			}
+
+			return clientCtx.PrintProto(res)
+		},
+	}
+
+	flags.AddQueryFlagsToCmd(cmd)
+
+	return cmd
+}
