+++ x/observer/types/genesis.go
@@ -10,9 +10,10 @@ import (
 func DefaultGenesis() *GenesisState {
 	params := DefaultParams()
 	return &GenesisState{
-		Params:    &params,
-		Ballots:   nil,
-		Observers: nil,
+		Params:          &params,
+		Ballots:         nil,
+		Observers:       nil,
+		NodeAccountList: []*NodeAccount{},
 	}
 }
 
@@ -24,6 +25,15 @@ func (gs GenesisState) Validate() error {
 		if err != nil {
 			return err
 		}
+	}
+	// Check for duplicated index in nodeAccount
+	nodeAccountIndexMap := make(map[string]bool)
+
+	for _, elem := range gs.NodeAccountList {
+		if _, ok := nodeAccountIndexMap[elem.GetOperator()]; ok {
+			return fmt.Errorf("duplicated index for nodeAccount")
+		}
+		nodeAccountIndexMap[elem.GetOperator()] = true
 	}
 	return nil
 }
